requirements.txt
openai==0.28
streamlit>=1.25.0
python-dotenv>=1.0.0
